requirements.txt
asyncio==3.4.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gspread==6.0.2
gspread_asyncio==2.0.0
idna==3.7
oauthlib==3.2.2
playwright==1.44.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyee==11.1.0
python-dotenv==1.0.1
requests==2.32.2
requests-oauthlib==2.0.0
rsa==4.9
StrEnum==0.4.15
typing_extensions==4.12.0
urllib3==2.2.1
